Add select/permute optimized for nn (cpu/gpu)

+++ include/eddl/hardware/gpu/gpu_kernels.h
@@ -26,11 +26,11 @@
 __global__ void fill_(float *A, float v, long int size);
 __global__ void fill(float *aptr, float *bptr, int t, int aini, int at, int bini, int bt, int tot, int inc);
 
-__global__ void select(float *A, float* B, int size, int* indices);
-__global__ void select_back(float *A, float* B, int size, int* indices);
+__global__ void select(float *A, float* B, long int size, int* indices);
+__global__ void select_back(float *A, float* B, long int size, int* indices);
 
-__global__ void set_select(float *A, float* B, int size, int* indices);
-__global__ void set_select_back(float *A, float* B, int size, int* indices);
+__global__ void set_select(float *A, float* B, long int size, int* indices);
+__global__ void set_select_back(float *A, float* B, long int size, int* indices);
 
 __global__ void select_rows(float *A, float* B, int rowsize, int size, int* indices, int ini,bool mask_zeros);  // TODO: Legacy
 __global__ void deselect_rows(float *A, float* B, int rowsize, int size, int* indices, int ini,int inc,bool mask_zeros);  // TODO: Legacy
